Added Delta values for trade stats

+++ MabiCommerce/Domain/Trading/Load.cs
@@ -32,13 +32,13 @@ namespace MabiCommerce.Domain.Trading
 			return _slots.Sum(kvp => kvp.Key.Profits.First(x => x.Destination == destination).Amount * kvp.Value);
 		}
 
-		public long CalculateExperience(TradingPost destination)
+		public int CalculateExperience(TradingPost destination)
 		{
 			// http://wiki.mabinogiworld.com/view/Commerce
 			// floor(sqrt(Single Item Profit * Single Item Weight)) * Item Quantity * 30
 
 			return _slots.Sum(kvp =>
-				(long)(Math.Sqrt(kvp.Key.Profits.First(x => x.Destination == destination).Amount * kvp.Key.Weight)) * kvp.Value * 30);
+				(int)(Math.Sqrt(kvp.Key.Profits.First(x => x.Destination == destination).Amount * kvp.Key.Weight)) * kvp.Value * 30);
 		}
 
 		public override string ToString()
